attempt to reduce structures
there is no reason to keep zero Value structs around. They could eat up
a lot of memory as the length of Vector grows or the Dim of a Matrix
increases. This is the first of more PRs to follow. Better encapsulation
by restricting access to certain underline data structs